pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "homework"
version = "0.1.0"
description = "Homework projects collection"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    # Vector Database
    "pymilvus>=2.3.6,<3.0.0",
    "neo4j>=5.15.0,<6.0.0",
    
    # LlamaIndex
    "llama-index>=0.9.30,<0.10.0",
    "llama-index-vector-stores-milvus>=0.1.11,<0.2.0",
    "llama-index-embeddings-openai>=0.1.7,<0.2.0",
    "llama-index-llms-openai>=0.1.7,<0.2.0",
    
    # FastAPI
    "fastapi>=0.104.0,<1.0.0",
    "uvicorn>=0.24.0,<1.0.0",
    "pydantic>=2.5.0,<3.0.0",
    "pydantic-settings>=2.0.0,<3.0.0",
    
    # Text Processing
    "sentence-transformers>=2.2.0,<3.0.0",
    "jieba>=0.42.0,<1.0.0",
    "nltk>=3.8.0,<4.0.0",
    
    # Utilities
    "python-dotenv>=1.0.0,<2.0.0",
    "loguru>=0.7.2,<1.0.0",
    "watchdog>=3.0.0,<4.0.0",
    "pandas>=2.1.0,<3.0.0",
    "numpy>=1.24.0,<2.0.0",
    
    # Development
    "pytest>=7.4.0,<8.0.0",
    "pytest-asyncio>=0.21.0,<1.0.0",
    "requests>=2.31.0,<3.0.0",
    
    # MultiJump dependencies
    "rich>=13.7.0,<14.0.0",
    "typer>=0.9.0,<1.0.0",
]

[project.optional-dependencies]
dev = [
    "black",
    "isort",
    "flake8",
    "mypy",
]

[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 88

[tool.setuptools.packages.find]
where = ["."]
include = ["week*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"


